python: Make DAR parsing compatible with bigger DALFs.
Note that separately, this PR also removes formal support for Python 3.6 because Python 3.6 has been end-of-lifed, and some of the build tools that dazl relies on no longer support Python 3.6. The changes in dazl 7.9.0 are minimal from 7.8.4, but it should be noted that 7.8.4 will have been the last version _tested_ against Python 3.6. You are strongly advised to upgrade to a supported version of Python, and note that dazl 8.0.0 will remove support for Python 3.6, as many other libraries have also removed support for Python 3.6.